// src/client/main/infra/apis/api-core/models/RegisterDocumentResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type RegisterDocumentResponse = {
  id: string;
  orgId: string;
  companyId: string;
  editalId: string;
  type: string;
  url: string;
  publishedAt: (string | null);
  createdAt: string;
};