// src/client/main/infra/apis/api-core/models/RunDocumentSummaryResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type RunDocumentSummaryResponse = {
  id: string;
  orgId: string;
  companyId: string;
  documentId: string;
  version: number;
  status: string;
  model: (string | null);
  summary: (string | null);
  createdAt: string;
  updatedAt: string;
};